proyecto-asistencia/servidor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python==4.9.0.80
face-recognition==1.3.0
mysql-connector-python==8.3.0
pillow==10.2.0
pydantic==2.5.3
requests==2.31.0
python-multipart==0.0.6
numpy==1.26.3